refactor: глобальные исправления в fonds и capital sources и в справочниках

+++ src/main/java/ru/pssbd/fonds/dto/input/CapitalSourceInput.java
@@ -30,6 +30,6 @@ public class CapitalSourceInput {
     private List<Integer> fonds;
 
     @NotNull
-    private List<Integer> donationTypes;
+    private List<Short> donationTypes;
 
 }
